@@ -13,12 +13,12 @@
     "@types/jest": "^28.1.6",
     "bintrees": "^1.0.2",
     "collections": "^5.1.12",
-    "csv-stringify": "^5.6.2",
-    "eslint": "^7.32.0",
-    "eslint-config-pipedrive": "^9.0.1",
+    "csv-stringify": "^6.2.0",
+    "eslint": "^8.20.0",
+    "eslint-config-pipedrive": "^13.0.0",
     "eslint-config-prettier": "^8.3.0",
     "functional-red-black-tree": "^1.0.1",
-    "jest": "^27.0.6",
+    "jest": "^28.1.3",
     "microtime": "^3.0.0",
     "prettier": "^2.3.2",
     "sorted-btree": "^1.5.0",
