@@ -1,9 +1,9 @@
 "use strict";
-exports.__esModule = true;
-var utils_1 = require("../src/utils/utils");
-describe("indexOf", function () {
-    describe("returns the expected positions", function () {
-        var arr = [];
+Object.defineProperty(exports, "__esModule", { value: true });
+const utils_1 = require("../src/utils/utils");
+describe("indexOf", () => {
+    describe("returns the expected positions", () => {
+        let arr = [];
         // if it is empty, then what is to be added should go to the zeroth position
         expect((0, utils_1.indexOf)(arr, arr.length, 1, utils_1.defaultComparator)).toEqual(0);
         arr = [1, 2, 5, 9];
@@ -19,19 +19,20 @@ describe("indexOf", function () {
         expect((0, utils_1.indexOf)(arr, arr.length, 0, utils_1.defaultComparator)).toEqual(0);
     });
 });
-describe("mostSignificantBit", function () {
+describe("mostSignificantBit", () => {
     describe('correctly returns the value of the most significant bit', function () {
-        var cases = [[1, 1], [11, 8], [219, 128], [249, 128], [255, 128], [256, 256]];
-        test.each(cases)("number %d expected msb value %d", function (value, msbValue) {
+        const cases = [[1, 1], [11, 8], [219, 128], [249, 128], [255, 128], [256, 256]];
+        test.each(cases)("number %d expected msb value %d", (value, msbValue) => {
             expect((0, utils_1.mostSignificantBit)(value)).toEqual(msbValue);
         });
     });
 });
-describe("leastSignificantBit", function () {
+describe("leastSignificantBit", () => {
     describe('correctly returns the value of the least significant bit', function () {
-        var cases = [[1, 1], [11, 1], [16, 16], [32, 32], [55, 1], [256, 256]];
-        test.each(cases)("number %d expected lsb value %d", function (value, lsbValue) {
+        const cases = [[1, 1], [11, 1], [16, 16], [32, 32], [55, 1], [256, 256]];
+        test.each(cases)("number %d expected lsb value %d", (value, lsbValue) => {
             expect((0, utils_1.leastSignificantBit)(value)).toEqual(lsbValue);
         });
     });
 });
+//# sourceMappingURL=utils.test.js.map
